refactor: rename `functions_loader_new`
